Log with selecta that pipes into git show, allows parameters.
A quite common use case for me is to look at commits and
check one of these in particular. With this alias this is
possible.

A second use case is to search for all commits related to a story and/or
author which can be done with 'git ss --grep XX-100' or git ss --author
mvaltas